Save query

+++ models/Models.py
@@ -28,7 +28,42 @@ class JsonModel:
 
 
 class BaseModel(Model):
-    def FillAndSave(self, data, save=True):
+    def save_query(self, force_insert=False, only=None):
+
+        query = ""
+        field_dict = self.__data__.copy()
+        if self._meta.primary_key is not False:
+            pk_field = self._meta.primary_key
+            pk_value = self._pk
+        else:
+            pk_field = pk_value = None
+        if only:
+            field_dict = self._prune_fields(field_dict, only)
+        elif self._meta.only_save_dirty and not force_insert:
+            field_dict = self._prune_fields(field_dict, self.dirty_fields)
+            if not field_dict:
+                self._dirty.clear()
+                return False
+
+        self._populate_unsaved_relations(field_dict)
+        if pk_value is not None and not force_insert:
+            if self._meta.composite_key:
+                for pk_part_name in pk_field.field_names:
+                    field_dict.pop(pk_part_name, None)
+            else:
+                field_dict.pop(pk_field.name, None)
+            query = self.update(**field_dict).where(self._pk_expr()).sql()
+        elif pk_field is None or not self._meta.auto_increment:
+            query = self.insert(**field_dict).sql()
+        else:
+            query = self.insert(**field_dict).sql()
+
+        result = self._meta.database.connection().cursor().mogrify(query[0],query[1])
+        return result
+
+
+    def GetQuery(self, data, save=True):
+        query = ()
         for key in self._definition.field_assoc:
             if (not hasattr(self, key)):
                 continue
@@ -48,7 +83,51 @@ class BaseModel(Model):
                     sub_model = attr
                     fk_field = getattr(sub_model, field_data.get("attributes").get("fk"))
                     sub_model.bind(self._meta.database, bind_refs=False, bind_backrefs=False)
-                    sub_model.delete().where(fk_field == item_id).execute()
+                    _query = sub_model.delete().where(fk_field == item_id).sql()
+
+                    query+= self._meta.database.connection().cursor().mogrify(_query[0],_query[1]),
+                    subfields = field_data["fields"]
+                    sub_model._definition.field_assoc = {subfields[i]["db_name"]: subfields[i]["name"] for i in
+                                                         subfields}
+                    for sub_data in data_val:
+                        sub_item = sub_model()
+                        #try:
+                        if True:
+                            setattr(sub_item, field_data.get("attributes").get("fk"), item_id)
+                            sub_item = sub_item.GetQuery(sub_data, False)
+                            query += sub_item
+                        #except Exception as e:
+                        #    print(e)
+                            pass
+
+        if save:
+            query += self.save_query(),
+        return query
+
+    def FillAndSave(self, data, save=True):
+        query = ()
+        for key in self._definition.field_assoc:
+            if (not hasattr(self, key)):
+                continue
+
+            val = self._definition.field_assoc[key]
+            if (not save):
+                self.owner_id = 1
+
+            data_val = data[val]
+            if (type(data_val) != list):
+                setattr(self, key, data[val])
+            else:
+                attr = getattr(self, key)
+                if (issubclass(attr, BaseModel)):
+                    item_id = getattr(self, self._definition.pk)
+                    field_data = self._definition.field_data[key]
+                    sub_model = attr
+                    fk_field = getattr(sub_model, field_data.get("attributes").get("fk"))
+                    sub_model.bind(self._meta.database, bind_refs=False, bind_backrefs=False)
+                    query=sub_model.delete().where(fk_field == item_id).execute()
+                    #print(_query)
+                    query+=_query
                     subfields = field_data["fields"]
                     sub_model._definition.field_assoc = {subfields[i]["db_name"]: subfields[i]["name"] for i in
                                                          subfields}
@@ -57,14 +136,14 @@ class BaseModel(Model):
                         try:
                             setattr(sub_item, field_data.get("attributes").get("fk"), item_id)
                             sub_item = sub_item.FillAndSave(sub_data, False)
-                            sub_item.save()
+                            query += sub_item
                         except Exception as e:
                             print(e)
                             pass
 
         if save:
             self.save()
-        return self
+        return query
 
     class Meta:
         pass
@@ -97,7 +176,7 @@ class Groups(BaseModel):
         chance = IntegerField()
         index = IntegerField()
 
-        def FillAndSave(self, data, save=True):
+        def SetData(self, data):
             _data = {
                 'index': data["index"],
                 'chance': data["chance"],
@@ -114,6 +193,14 @@ class Groups(BaseModel):
                 _data['chest_group_id'] = data["item_id"]
                 self.chest_group_id = data["item_id"]
 
+
+        def GetQuery(self, data, save=True):
+            self.SetData(data)
+            return super(Groups.chest_contents, self).GetQuery(data)
+
+
+        def FillAndSave(self, data, save=True):
+            self.SetData(data)
             return super(Groups.chest_contents, self).FillAndSave(data)
 
         class _definition:
